third_party/WebKit/LayoutTests/bluetooth/service/getCharacteristics/gen-reconnect-during-with-uuid.html

<!-- Generated by //third_party/WebKit/LayoutTests/bluetooth/generate.py -->
<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/testdriver.js"></script>
<script src="../../../resources/testdriver-vendor.js"></script>
<script src="../../../external/wpt/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
'use strict';
bluetooth_test(() => {
  let device, service;
  return getHealthThermometerDeviceWithServicesDiscovered({
    filters: [{services: [health_thermometer.name]}],
  })
    .then(_ => ({device} = _))
    .then(() => device.gatt.getPrimaryService(health_thermometer.name))
    .then(_ => service = _)
    .then(() => {
      // 1. Make a call to service.getCharacteristics, while the service is still
      // valid.
      let promise = assert_promise_rejects_with_message(service.getCharacteristics(measurement_interval.name), new DOMException(
          'GATT Server is disconnected. Cannot retrieve characteristics. ' +
          '(Re)connect first with `device.gatt.connect`.',
          'NetworkError'));

      // 2. disconnect() and connect before the initial call completes.
      // This is accomplished by making the calls without waiting for the
      // earlier promises to resolve.
      // connect() guarantees on OS-level connection, but disconnect()
      // only disconnects the current instance.
      // getHealthThermometerDeviceWithServicesDiscovered holds another
      // connection in an iframe, so disconnect() and connect() are certain to
      // reconnect.  However, disconnect() will invalidate the service object so
      // the subsequent calls made to it will fail, even after reconnecting.
      device.gatt.disconnect();
      return device.gatt.connect()
        // 3. Check that the initial call did eventually fail.  To do that, we
        // must hold a reference to the first promise.
        .then(() => promise);
    });
}, 'disconnect() and connect() called during getCharacteristics. Reject with ' +
   'NetworkError.');

</script>
